project_2_memeGenerator/requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.2
MarkupSafe==1.1.1
numpy==1.20.1
pandas==1.2.2
Pillow==8.1.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2021.1
six==1.15.0
Werkzeug==1.0.1
